Anka/school/utils/requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dlib==19.17.0
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.2.3
face-recognition-models==0.3.0
flake8==3.7.7
graphviz==0.8.4
grpcio==1.21.1
idna==2.8
jdcal==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mxnet==1.3.1
numpy==1.14.6
opencv-python==4.0.0.21
openpyxl==2.6.2
Pillow==5.4.1
protobuf==3.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
requests==2.22.0
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tqdm==4.32.2
urllib3==1.25.3
Werkzeug==0.15.4
